add User management and auth to the app

+++ app/Http/Controllers/ListingController.php
@@ -39,9 +39,11 @@ class ListingController extends Controller
             'description' => 'required',
         ]);
 
-        if($request->hasFile('logo')){
+        if($request->hasFile('logo')) {
             $formFields['logo'] = $request->file('logo')->store('logos', 'public');
         }
+
+        $formFields['user_id'] = auth()->id();
 
         Listing::create($formFields);
         
@@ -55,6 +57,11 @@ class ListingController extends Controller
 
     // Update Listing Data
     public function update(Request $request, Listing $listing){
+        // Make sure logged User is owner
+        if($listing->user_id != auth()->id()){
+            abort(403, 'Unauthorized Action!');
+        }
+
         $formFields = $request->validate([
             'title' => 'required',
             'company' => ['required'],
@@ -76,7 +83,17 @@ class ListingController extends Controller
 
     // Delete Listing
     public function destroy(Listing $listing){
+        // Make sure logged User is owner
+        if($listing->user_id != auth()->id()){
+            abort(403, 'Unauthorized Action!');
+        }
+        
         $listing->delete();
         return redirect('/')->with('message', 'listing deleted successfully');
+    }
+
+    // Manage Listings
+    public function manage(){
+        return view('listings.manage', ['listings' => auth()->user()->listings()->get()]);
     }
 }
